Backend/requirements.txt
azure-common==1.1.28
azure-core==1.31.0
azure-cosmos==4.7.0
azure-identity==1.19.0
azure-mgmt-core==1.4.0
azure-mgmt-web==7.3.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dotenv-python==0.0.1
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
idna==3.7
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.31.0
msal-extensions==1.2.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scipy==1.14.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
